Remove redundant backend code and improve unit tests
