--- my-react-app/Testing/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dummy Login Page</title>
-</head>
-<body>
-    <h1>Dummy Login Page</h1>
-    <form id="loginForm">
-        <label for="employeeID">Employee ID:</label>
-        <input type="text" id="employeeID" name="employeeID" value="2">
-        <button type="submit">Login</button>
-    </form>
-
-    <script src="script.js"></script>
-</body>
-</html>
