--- my-react-app/Testing/loggedin.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Logged In</title>
-</head>
-<body>
-    <h1>You are logged in!</h1>
-</body>
-</html>
