+++ my-react-app/package.json
@@ -6,8 +6,13 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.6.8",
+    "body-parser": "^1.20.2",
     "bootstrap": "^5.3.3",
+    "cors": "^2.8.5",
+    "express": "^4.19.2",
     "jest": "^27.5.1",
+    "mssql": "^10.0.2",
     "nth-check": "^2.1.1",
     "postcss": "^8.4.38",
     "range": "^0.0.3",
