+++ package.json
@@ -7,12 +7,16 @@
     "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "chai": "^5.1.0",
+    "mocha": "^10.4.0"
   },
   "name": "zeit3118",
   "version": "1.0.0",
   "main": "NodeTest.js",
-  "scripts": {},
+  "scripts": {
+    "test": "mocha"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/jayathg/ZEIT3118.git"
